refactor: Common Module config package refactor (#10)
